use -Wconversion and -Wsign-conversion and fix any errors reported

+++ src/main.c
@@ -8,7 +8,7 @@
 int main(int argc, char **argv)
 {
     if (argc > 1 || !isatty(STDIN_FILENO))
-        return ncsh_noninteractive(argc, argv);
+        return (int)ncsh_noninteractive(argc, argv);
     else
-        return ncsh();
+        return (int)ncsh();
 }
